Factor out getRenameInfo
